requirements.txt
ipython
Django==5.1.4
djangorestframework
django-cors-headers
openai
requests
python-dotenv
unstructured[all-docs]
unstructured-ingest
langchain-openai
unstructured[openai]